+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 * Fixed native memory leak setting the value of a primary key (#3993).
 * Activated Realm's annotation processor on connectedTest when the project is using kapt (#4008).
+* Fixed bug, preventing Sync client to renew the access token (#4038) (#4039).
 
 ### Object Server API Changes (In Beta)
 
@@ -19,6 +20,7 @@
 ### Internal
 
 * Updated Realm Sync to 1.0.0-BETA-7.1.
+* Add a Realm backup when receiving a Sync client reset message from the server.
 
 ## 2.2.2